support backend sampling for sessions
