+++ backend/private-graph/graph/schema.graphqls
@@ -24,6 +24,8 @@ enum SessionExcludedReason {
 	IgnoredUser
 	BillingQuotaExceeded
 	RetentionPeriodExceeded
+	Sampled
+	ExclusionFilter
 }
 
 type Session {
