+++ backend/public-graph/graph/resolver.go
@@ -3287,80 +3287,6 @@ func (r *Resolver) SendSessionUserPropertiesAlert(ctx context.Context, workspace
 	return nil
 }
 
-func (r *Resolver) isSessionExcluded(ctx context.Context, s *model.Session, sessionHasErrors bool) (bool, *privateModel.SessionExcludedReason) {
-	var excluded bool
-	var reason privateModel.SessionExcludedReason
-
-	var project model.Project
-	if err := r.DB.Raw("SELECT * FROM projects WHERE id = ?;", s.ProjectID).Scan(&project).Error; err != nil {
-		log.WithContext(ctx).WithFields(log.Fields{"session_id": s.ID, "project_id": s.ProjectID, "identifier": s.Identifier}).Errorf("error fetching project for session: %v", err)
-		return false, nil
-	}
-
-	if r.isSessionUserExcluded(ctx, s, project) {
-		excluded = true
-		reason = privateModel.SessionExcludedReasonIgnoredUser
-	}
-
-	if r.isSessionExcludedForNoError(ctx, s, &project, sessionHasErrors) {
-		excluded = true
-		reason = privateModel.SessionExcludedReasonNoError
-	}
-
-	if r.isSessionExcludedForNoUserEvents(ctx, s) {
-		excluded = true
-		reason = privateModel.SessionExcludedReasonNoUserEvents
-	}
-
-	return excluded, &reason
-}
-
-func (r *Resolver) isSessionExcludedForNoUserEvents(ctx context.Context, s *model.Session) bool {
-	return s.LastUserInteractionTime.Unix() == 0
-}
-
-func (r *Resolver) isSessionExcludedForNoError(ctx context.Context, s *model.Session, project *model.Project, sessionHasErrors bool) bool {
-	projectFilterSettings, _ := r.Store.GetProjectFilterSettings(ctx, project.ID)
-
-	if projectFilterSettings.FilterSessionsWithoutError {
-		return !sessionHasErrors
-	}
-
-	return false
-}
-
-func (r *Resolver) isSessionUserExcluded(ctx context.Context, s *model.Session, project model.Project) bool {
-	if project.ExcludedUsers == nil {
-		return false
-	}
-	var email string
-	if s.UserProperties != "" {
-		encodedProperties := []byte(s.UserProperties)
-		decodedProperties := map[string]string{}
-		err := json.Unmarshal(encodedProperties, &decodedProperties)
-		if err != nil {
-			log.WithContext(ctx).WithFields(log.Fields{"session_id": s.ID, "project_id": s.ProjectID}).Errorf("Could not unmarshal user properties: %s, error: %v", s.UserProperties, err)
-			return false
-		}
-		email = decodedProperties["email"]
-	}
-	for _, value := range []string{s.Identifier, email} {
-		if value == "" {
-			continue
-		}
-		for _, excludedExpr := range project.ExcludedUsers {
-			matched, err := regexp.MatchString(excludedExpr, value)
-			if err != nil {
-				log.WithContext(ctx).WithFields(log.Fields{"session_id": s.ID, "project_id": s.ProjectID}).Errorf("error running regexp for excluded users: %s with value: %s, error: %v", excludedExpr, value, err.Error())
-				return false
-			} else if matched {
-				return true
-			}
-		}
-	}
-	return false
-}
-
 func (r *Resolver) isExcludedError(ctx context.Context, errorFilters []string, errorEvent string, projectID int) bool {
 	if errorEvent == "[{}]" {
 		log.WithContext(ctx).
